📁 organized components into folders + ✨ added sections for the landing page
