+++ src/Components/LandingSection/LandingSection.css
@@ -1,0 +1,21 @@
+#home-section {
+    height: 100vh;
+    width: 100%;
+    display: grid;
+    /*align-items: center;
+    justify-content: center;*/
+    background-size: cover;
+    
+}
+
+.red {
+    background: red no-repeat center center fixed;
+}
+
+.blue {
+    background: blue no-repeat center center fixed;
+}
+
+h1 {
+    text-align: center;
+}
